pyproject.toml
[project]
name = "bambushot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bambu-connect>=0.3.1",
    "einops>=0.8.0",
    "ipdb>=0.13.13",
    "matplotlib>=3.10.0",
    "numpy>=2.2.1",
    "opencv-python>=4.10.0.84",
    "pillow>=11.0.0",
    "python-dotenv>=1.0.1",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "transformers>=4.47.1",
]
